Refactored parameter tests.
