+++ glotaran/parameter/test/test_parameter_group.py
@@ -1,106 +1,242 @@
-from IPython.core.formatters import format_display_data
+from __future__ import annotations
+
+import numpy as np
 
 from glotaran.io import load_parameters
-from glotaran.parameter.parameter_group import ParameterGroup
-
-PARAMETERS_3C_BASE = """\
-irf:
-    - ["center", 1.3]
-    - ["width", 7.8]
-j:
-    - ["1", 1, {"vary": False, "non-negative": False}]
-"""
-
-PARAMETERS_3C_KINETIC = """\
-kinetic:
-    - ["1", 300e-3]
-    - ["2", 500e-4]
-    - ["3", 700e-5]
-"""
-
-RENDERED_MARKDOWN = """\
-  * __irf__:
-
-    | _Label_   |   _Value_ |   _StdErr_ |   _Min_ |   _Max_ | _Vary_   | _Non-Negative_   | _Expr_   |
-    |-----------|-----------|------------|---------|---------|----------|------------------|----------|
-    | center    |       1.3 |          0 |    -inf |     inf | True     | False            | None     |
-    | width     |       7.8 |          0 |    -inf |     inf | True     | False            | None     |
-
-  * __j__:
-
-    |   _Label_ |   _Value_ |   _StdErr_ |   _Min_ |   _Max_ | _Vary_   | _Non-Negative_   | _Expr_   |
-    |-----------|-----------|------------|---------|---------|----------|------------------|----------|
-    |         1 |         1 |          0 |    -inf |     inf | False    | False            | None     |
-
-  * __kinetic__:
-
-    |   _Label_ |   _Value_ |   _StdErr_ |   _Min_ |   _Max_ | _Vary_   | _Non-Negative_   | _Expr_   |
-    |-----------|-----------|------------|---------|---------|----------|------------------|----------|
-    |         1 |     0.3   |          0 |    -inf |     inf | True     | False            | None     |
-    |         2 |     0.05  |          0 |    -inf |     inf | True     | False            | None     |
-    |         3 |     0.007 |          0 |    -inf |     inf | True     | False            | None     |
-
-"""  # noqa: E501
+from glotaran.io import save_parameters
 
 
-def test_param_group_markdown_is_order_independent():
-    """Markdown output of ParameterGroup.markdown() is independent of initial order"""
-    PARAMETERS_3C_INITIAL1 = f"""{PARAMETERS_3C_BASE}\n{PARAMETERS_3C_KINETIC}"""
-    PARAMETERS_3C_INITIAL2 = f"""{PARAMETERS_3C_KINETIC}\n{PARAMETERS_3C_BASE}"""
+def test_parameter_group_copy():
+    params = """
+    kinetic:
+        - ["5", 1, {non-negative: true, min: -1, max: 1, vary: false}]
+        - 4
+        - 5
+    j:
+        - 7
+        - 8
+    """
+    params = load_parameters(params, format_name="yml_str")
+    copy = params.copy()
 
-    initial_parameters_ref = ParameterGroup.from_dict(
-        {
-            "j": [["1", 1, {"vary": False, "non-negative": False}]],
-            "kinetic": [
-                ["1", 300e-3],
-                ["2", 500e-4],
-                ["3", 700e-5],
-            ],
-            "irf": [["center", 1.3], ["width", 7.8]],
-        }
+    for label, parameter in params.all():
+        assert copy.has(label)
+        copied_parameter = copy.get(label)
+        assert parameter.value == copied_parameter.value
+        assert parameter.non_negative == copied_parameter.non_negative
+        assert parameter.minimum == copied_parameter.minimum
+        assert parameter.maximum == copied_parameter.maximum
+        assert parameter.vary == copied_parameter.vary
+
+
+def test_parameter_group_from_list():
+    params = """
+    - 5
+    - 4
+    - 3
+    - 2
+    - 1
+    """
+
+    params = load_parameters(params, format_name="yml_str")
+
+    assert len(list(params.all())) == 5
+
+    assert [p.label for _, p in params.all()] == [f"{i}" for i in range(1, 6)]
+    assert [p.value for _, p in params.all()] == list(range(1, 6))[::-1]
+
+
+def test_parameter_group_from_dict():
+    params = """
+    kinetic:
+        - 3
+        - 4
+        - 5
+    j:
+        - 7
+        - 8
+    """
+
+    params = load_parameters(params, format_name="yml_str")
+
+    assert len(list(params.all())) == 5
+    group = params["kinetic"]
+    assert len(list(group.all())) == 3
+    assert [p.label for _, p in group.all()] == [f"{i}" for i in range(1, 4)]
+    assert [p.value for _, p in group.all()] == list(range(3, 6))
+    group = params["j"]
+    assert len(list(group.all())) == 2
+    assert [p.label for _, p in group.all()] == [f"{i}" for i in range(1, 3)]
+    assert [p.value for _, p in group.all()] == list(range(7, 9))
+
+
+def test_parameter_group_from_dict_nested():
+    params = """
+    kinetic:
+        j:
+            - 7
+            - 8
+            - 9
+    """
+
+    params = load_parameters(params, format_name="yml_str")
+    assert len(list(params.all())) == 3
+    group = params["kinetic"]
+    assert len(list(group.all())) == 3
+    group = group["j"]
+    assert len(list(group.all())) == 3
+    assert [p.label for _, p in group.all()] == [f"{i}" for i in range(1, 4)]
+    assert [p.value for _, p in group.all()] == list(range(7, 10))
+
+
+def test_parameter_group_to_array():
+    params = """
+    - ["1", 1, {non-negative: false, min: -1, max: 1, vary: false}]
+    - ["2", 4e2, {non-negative: true, min: 10, max: 8e2, vary: true}]
+    - ["3", 2e4]
+    """
+
+    params = load_parameters(params, format_name="yml_str")
+
+    labels, values, lower_bounds, upper_bounds = params.get_label_value_and_bounds_arrays(
+        exclude_non_vary=False
     )
 
-    initial_parameters1 = load_parameters(PARAMETERS_3C_INITIAL1, format_name="yml_str")
-    initial_parameters2 = load_parameters(PARAMETERS_3C_INITIAL2, format_name="yml_str")
+    assert len(labels) == 3
+    assert len(values) == 3
+    assert len(lower_bounds) == 3
+    assert len(upper_bounds) == 3
 
-    assert initial_parameters1.markdown() == RENDERED_MARKDOWN
-    assert initial_parameters2.markdown() == RENDERED_MARKDOWN
-    assert initial_parameters_ref.markdown() == RENDERED_MARKDOWN
+    assert labels == ["1", "2", "3"]
+    assert np.allclose(values, [1, np.log(4e2), 2e4])
+    assert np.allclose(lower_bounds, [-1, np.log(10), -np.inf])
+    assert np.allclose(upper_bounds, [1, np.log(8e2), np.inf])
 
+    (
+        labels_only_vary,
+        values_only_vary,
+        lower_bounds_only_vary,
+        upper_bounds_only_vary,
+    ) = params.get_label_value_and_bounds_arrays(exclude_non_vary=True)
 
-def test_param_group_repr():
-    """Repr creates code to recreate the object with from_dict."""
-    result = ParameterGroup.from_dict({"foo": {"bar": [["1", 1.0], ["2", 2.0], ["3", 3.0]]}})
-    result_short = ParameterGroup.from_dict({"foo": {"bar": [1, 2, 3]}})
-    expected = "ParameterGroup.from_dict({'foo': {'bar': [['1', 1.0], ['2', 2.0], ['3', 3.0]]}})"
+    assert len(labels_only_vary) == 2
+    assert len(values_only_vary) == 2
+    assert len(lower_bounds_only_vary) == 2
+    assert len(upper_bounds_only_vary) == 2
 
-    assert result == result_short
-    assert result_short.__repr__() == expected
-    assert result.__repr__() == expected
-    assert result == eval(result.__repr__())
-
-
-def test_param_group_repr_from_list():
-    """Repr creates code to recreate the object with from_list."""
-    result = ParameterGroup.from_list([["1", 2.3], ["2", 3.0]])
-    result_short = ParameterGroup.from_list([2.3, 3.0])
-    expected = "ParameterGroup.from_list([['1', 2.3], ['2', 3.0]])"
-
-    assert result == result_short
-    assert result.__repr__() == expected
-    assert result_short.__repr__() == expected
-    assert result == eval(result.__repr__())
+    assert labels_only_vary == ["2", "3"]
 
 
-def test_param_group_ipython_rendering():
-    """Autorendering in ipython"""
-    param_group = ParameterGroup.from_dict({"foo": {"bar": [["1", 1.0], ["2", 2.0], ["3", 3.0]]}})
-    rendered_obj = format_display_data(param_group)[0]
+def test_parameter_group_set_from_label_and_value_arrays():
+    params = """
+    - ["1", 1, {non-negative: false, min: -1, max: 1, vary: false}]
+    - ["2", 4e2, {non-negative: true, min: 10, max: 8e2, vary: true}]
+    - ["3", 2e4]
+    """
 
-    assert "text/markdown" in rendered_obj
-    assert rendered_obj["text/markdown"].startswith("  * __foo__")
+    params = load_parameters(params, format_name="yml_str")
 
-    rendered_markdown_return = format_display_data(param_group.markdown())[0]
+    labels = ["1", "2", "3"]
+    values = [0, np.log(6e2), 42]
 
-    assert "text/markdown" in rendered_markdown_return
-    assert rendered_markdown_return["text/markdown"].startswith("  * __foo__")
+    params.set_from_label_and_value_arrays(labels, values)
+
+    values[1] = np.exp(values[1])
+
+    for i in range(3):
+        assert params.get(f"{i+1}").value == values[i]
+
+
+def test_parameter_group_from_csv(tmpdir):
+
+    TEST_CSV = """
+label, value, minimum, maximum, vary, non-negative, expression
+rates.k1,0.050,0,5,True,True,None
+rates.k2,None,-inf,inf,True,True,$rates.k1 * 2
+rates.k3,2.311,-inf,inf,True,True,None
+pen.eq.1,1.000,-inf,inf,False,False,None
+"""
+
+    csv_path = tmpdir.join("parameters.csv")
+    with open(csv_path, "w") as f:
+        f.write(TEST_CSV)
+
+    params = load_parameters(csv_path)
+
+    assert "rates" in params
+
+    assert params.has("rates.k1")
+    p = params.get("rates.k1")
+    assert p.label == "k1"
+    assert p.value == 0.05
+    assert p.minimum == 0
+    assert p.maximum == 5
+    assert p.vary
+    assert p.non_negative
+    assert p.expression is None
+
+    assert params.has("rates.k2")
+    p = params.get("rates.k2")
+    assert p.label == "k2"
+    assert p.value == params.get("rates.k1") * 2
+    assert p.minimum == -np.inf
+    assert p.maximum == np.inf
+    assert not p.vary
+    assert not p.non_negative
+    assert p.expression == "$rates.k1 * 2"
+
+    assert params.has("rates.k3")
+    p = params.get("rates.k3")
+    assert p.label == "k3"
+    assert p.value == 2.311
+    assert p.minimum == -np.inf
+    assert p.maximum == np.inf
+    assert p.vary
+    assert p.non_negative
+    assert p.expression is None
+
+    assert "pen" in params
+    assert "eq" in params["pen"]
+
+    assert params.has("pen.eq.1")
+    p = params.get("pen.eq.1")
+    assert p.label == "1"
+    assert p.value == 1.0
+    assert p.minimum == -np.inf
+    assert p.maximum == np.inf
+    assert not p.vary
+    assert not p.non_negative
+    assert p.expression is None
+
+
+def test_parameter_group_to_csv(tmpdir):
+    csv_path = tmpdir.join("parameters.csv")
+    params = load_parameters(
+        """
+    b:
+        - ["1", 0.25, {vary: false, min: 0, max: 8}]
+        - ["2", 0.75, {expr: '1 - $b.1', non-negative: true}]
+    rates:
+        - ["total", 2]
+        - ["branch1", {expr: '$rates.total * $b.1'}]
+        - ["branch2", {expr: '$rates.total * $b.2'}]
+    """,
+        format_name="yml_str",
+    )
+
+    save_parameters(params, csv_path, "csv")
+
+    with open(csv_path) as f:
+        print(f.read())
+    params_from_csv = load_parameters(csv_path)
+
+    for label, p in params.all():
+        assert params_from_csv.has(label)
+        p_from_csv = params_from_csv.get(label)
+        assert p.label == p_from_csv.label
+        assert p.value == p_from_csv.value
+        assert p.minimum == p_from_csv.minimum
+        assert p.maximum == p_from_csv.maximum
+        assert p.vary == p_from_csv.vary
+        assert p.non_negative == p_from_csv.non_negative
+        assert p.expression == p_from_csv.expression
